+++ src/fldeff_cut.c
@@ -165,7 +165,7 @@ bool8 SetUpFieldMove_Cut(void)
     s16 x, y;
     u8 i, j;
     u8 tileBehavior;
-    u8 userAbility;
+    u16 userAbility;
     bool8 cutTiles[CUT_NORMAL_AREA];
     bool8 ret;
 
@@ -321,7 +321,7 @@ bool8 SetUpFieldMove_Cut_grass(void)
     s16 x, y;
     u8 i, j;
     u8 tileBehavior;
-    u8 userAbility;
+    u16 userAbility;
     bool8 cutTiles[CUT_NORMAL_AREA];
     bool8 ret;
